Move TestTarget to TestAsset directory for CS
